devop: make sure network fee is still working for swap
