+++ src/modules/swap/ModuleSwap.vue
@@ -1,12 +1,5 @@
 <template>
   <div class="mew-component--swap">
-    <app-network-settings-modal
-      :open-settings="openSettings"
-      :close="closeGasPrice"
-      :gas-price-modal="gasPriceModal"
-      @onLocalGasPrice="handleLocalGasPrice"
-      @close="closeGasPrice"
-    />
     <mew6-white-sheet>
       <mew-module
         :has-elevation="true"
@@ -223,16 +216,16 @@
              Swap Fee
             =====================================================================================
           -->
-          <swap-fee
+          <app-network-fee
             v-if="step > 0"
             :show-fee="showSwapFee"
             :getting-fee="loadingFee"
             :error="feeError"
             :total-fees="totalFees"
-            :open-gas-price-modal="openGasPriceModal"
             :gas-price-type="localGasType"
             :message="feeError"
             :not-enough-eth="notEnoughEth"
+            @onLocalGasPrice="handleLocalGasPrice"
           />
           <div class="text-center">
             <mew-button
@@ -256,14 +249,12 @@ import AppUserMsgBlock from '@/core/components/AppUserMsgBlock';
 import ModuleAddressBook from '@/modules/address-book/ModuleAddressBook';
 import SwapIcon from '@/assets/images/icons/icon-swap.svg';
 import SwapProvidersList from './components/SwapProvidersList.vue';
-import SwapFee from './components/SwapFee.vue';
-import AppNetworkSettingsModal from '@/core/components/AppNetworkSettingsModal.vue';
+import AppNetworkFee from '@/core/components/AppNetworkFee.vue';
 import Swapper from './handlers/handlerSwap';
 import { toBN, fromWei, toWei, _ } from 'web3-utils';
 import { mapGetters, mapState, mapActions } from 'vuex';
 import Notification from '@/modules/notifications/handlers/handlerNotification';
 import BigNumber from 'bignumber.js';
-import { EventBus } from '@/core/plugins/eventBus';
 import { Toast, WARNING } from '../toast/handler/handlerToast';
 import {
   TRENDING_SYMBOLS,
@@ -293,8 +284,7 @@ export default {
     AppUserMsgBlock,
     ModuleAddressBook,
     SwapProvidersList,
-    SwapFee,
-    AppNetworkSettingsModal
+    AppNetworkFee
   },
   props: {
     fromToken: {
@@ -364,7 +354,6 @@ export default {
       swapIcon: SwapIcon,
       isLoadingProviders: false,
       addressValue: {},
-      gasPriceModal: false,
       selectedProvider: {},
       localGasPrice: '0',
       localGasType: 'economy',
@@ -809,9 +798,6 @@ export default {
         this.setSwapTokens(tokens);
       }
     },
-    openGasPriceModal() {
-      this.gasPriceModal = true;
-    },
     setDefaults() {
       setImmediate(() => {
         this.fromTokenType =
@@ -1044,13 +1030,6 @@ export default {
         this.feeError = this.isFromTokenEth ? ethError : message;
       }
     },
-    openSettings() {
-      EventBus.$emit('toggleSettings');
-      this.gasPriceModal = false;
-    },
-    closeGasPrice() {
-      this.gasPriceModal = false;
-    },
     setWrappedBtc(symbol) {
       const foundToken = this.toTokens.find(
         item => item.symbol.toLowerCase() === symbol.toLowerCase()
@@ -1058,6 +1037,7 @@ export default {
       this.setToToken(foundToken);
     },
     handleLocalGasPrice(e) {
+      console.log(e);
       this.localGasPrice = e.gasPrice;
       this.localGasType = e.gasType;
       if (!_.isEmpty(this.currentTrade)) {
